requirements.txt
Django>=5.1.7
python-dotenv>=1.0.0
djangorestframework>=3.14.0